# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register/email": {
      "post": {
        "operationId": "AuthController_registerEmail_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Step 1: Start registration with email",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register/details": {
      "post": {
        "operationId": "AuthController_registerDetails_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDetailsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Step 2: Provide personal details",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register/password": {
      "post": {
        "operationId": "AuthController_registerPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Step 3: Set password and create account",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register/verify": {
      "post": {
        "operationId": "AuthController_verifyEmail_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Step 4: Verify email with OTP code",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register/resend-otp": {
      "post": {
        "operationId": "AuthController_resendOtp_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Resend email verification OTP",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register/seller/email": {
      "post": {
        "operationId": "AuthController_registerSellerEmail_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Seller Step 1: Start registration with email",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register/seller/profile": {
      "post": {
        "operationId": "AuthController_registerSellerProfile_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSellerDetailsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Seller Step 2: Provide personal + business details. Bank info is collected separately via POST /sellers/me/banking after login - do not send it here.",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register/seller/password": {
      "post": {
        "operationId": "AuthController_registerSellerPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSellerPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Seller Step 3: Set password and create account. Bank info is collected separately via POST /sellers/me/banking after login.",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register/seller/verify": {
      "post": {
        "operationId": "AuthController_verifySellerEmail_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Seller Step 4: Verify email with OTP code",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created. Email a verification OTP is sent automatically. Caller should then hit /auth/otp/verify or /auth/register/verify.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Account created. Please verify your email.",
                    "email": "jane@example.com"
                  }
                }
              }
            }
          }
        },
        "summary": "Single-call registration (orchestrates email + details + password steps)",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/otp/request": {
      "post": {
        "operationId": "AuthController_otpRequest_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Channel is always email (Resend) regardless of \"phone\" wording in the PM spec. 60-second cooldown enforced by OtpService.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Login code sent"
                  }
                }
              }
            }
          }
        },
        "summary": "Send OTP via email for login/email-verification/password-reset",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_otpVerify_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response shape depends on the `purpose` field on the request.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "description": "purpose=login → returns JWT pair + user",
                      "example": {
                        "purpose": "login",
                        "tokens": {
                          "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                          "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                          "expiresIn": "15m"
                        },
                        "user": {
                          "id": "6710abc123def456789012ab",
                          "email": "jane@example.com",
                          "accountType": "BUYER"
                        }
                      }
                    },
                    {
                      "description": "purpose=email_verification → success message",
                      "example": {
                        "purpose": "email_verification",
                        "message": "Email verified successfully"
                      }
                    },
                    {
                      "description": "purpose=password_reset → short-lived resetToken",
                      "example": {
                        "purpose": "password_reset",
                        "resetToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Verify OTP code. Returns JWTs for login, message for verification, resetToken for password-reset",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/otp/resend": {
      "post": {
        "operationId": "AuthController_otpResend_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpResendDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Login code sent"
                  }
                }
              }
            }
          }
        },
        "summary": "Resend OTP (cooldown enforced)",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/clerk/login": {
      "post": {
        "operationId": "AuthController_clerkLogin_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClerkLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Login or register via Clerk OAuth (Google/Facebook)",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns JWT pair + user, or a requires2FA flag if 2FA is enabled.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                        "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                        "expiresIn": "15m",
                        "user": {
                          "id": "6710abc123def456789012ab",
                          "firstName": "Jane",
                          "lastName": "Doe",
                          "email": "jane@example.com",
                          "accountType": "BUYER",
                          "isVerified": true,
                          "isTwoFactorEnabled": false
                        }
                      }
                    },
                    {
                      "example": {
                        "requires2FA": true,
                        "email": "jane@example.com"
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or unverified email."
          }
        },
        "summary": "Login with email and password. For passwordless email-OTP login, use /auth/otp/request then /auth/otp/verify.",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login/password": {
      "post": {
        "operationId": "AuthController_loginPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Login with email and password (spec alias of /login)",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login/verify-2fa": {
      "post": {
        "operationId": "AuthController_verifyLogin2fa_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Verify2faLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Verify 2FA code during login",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Request password reset OTP",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/password/reset/request": {
      "post": {
        "operationId": "AuthController_passwordResetRequest_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Request password reset OTP (spec alias of /forgot-password)",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/forgot-password/verify": {
      "post": {
        "operationId": "AuthController_forgotPasswordVerify_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Verify password reset OTP",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/forgot-password/reset": {
      "post": {
        "operationId": "AuthController_forgotPasswordReset_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Reset password with reset token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/password/reset/confirm": {
      "post": {
        "operationId": "AuthController_passwordResetConfirm_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Confirm password reset with token (spec alias of /forgot-password/reset)",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/2fa/enable": {
      "post": {
        "operationId": "AuthController_enable2fa_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initiate 2FA enable (sends OTP)",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cadna Mart API",
    "description": "Backend API for the Cadna Mart marketplace.\n\n### Response envelope\nEvery successful response is wrapped by the global response interceptor:\n```json\n{\n  \"success\": true,\n  \"statusCode\": 200,\n  \"data\": { ... },\n  \"message\": \"...\",\n  \"meta\": { \"correlationId\": \"...\", \"timestamp\": \"...\" }\n}\n```\nWhen documenting individual endpoints below, the schema describes the `data` payload only.\n\n### Authentication\n- **Bearer JWT** (`Authorization: Bearer <accessToken>`) for authenticated user routes.\n- **Guest token** (`x-guest-token: <token>`) for guest-cart endpoints. Returned once from `POST /cart`.\n\n### Error shape\nErrors flow through the global `AllExceptionsFilter`:\n```json\n{\n  \"success\": false,\n  \"statusCode\": 400,\n  \"message\": \"Validation failed\",\n  \"errorCode\": \"BAD_REQUEST\",\n  \"path\": \"/api/v1/...\",\n  \"correlationId\": \"...\"\n}\n```",
    "version": "1.0",
    "contact": {
      "name": "Cadna Mart Engineering",
      "url": "",
      "email": "cadnamart@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "https://cadna-mart-be-nsz2.onrender.com/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned from /auth/login or /auth/otp/verify"
      },
      "guestToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-guest-token",
        "description": "Guest cart access token (returned once from POST /cart for guests)"
      }
    },
...
  }
}